requirements.txt
certifi==2019.3.9
chardet==3.0.4
coreapi==2.3.3
coreschema==0.0.4
Django==2.2.2
djangorestframework==3.9.4
drf-yasg==1.15.0
idna==2.8
inflection==0.3.1
itypes==1.1.0
Jinja2==2.10.1
MarkupSafe==1.1.1
pytz==2019.1
requests==2.22.0
ruamel.yaml==0.15.96
six==1.12.0
sqlparse==0.3.0
uritemplate==3.0.0
urllib3==1.25.3
